Refactor issue workflow CLI
